+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Updated dependencies
+- Inline code style in editor
 
 ## [0.8.124] - 2022-08-08
 ### Changed: